migrating the statistics to use the float for some averages rather than long (some stats are still not supported in the jcache-ehcache, and statistics support is not tested comprehensively yet in the tck so these unsupported stats slip through)

+++ ehcache-jcache/src/main/java/net/sf/ehcache/jcache/JCacheStatistics.java
@@ -186,8 +186,8 @@ public class JCacheStatistics implements javax.cache.CacheStatistics, Serializab
      * @return the time in milliseconds
      */
     @Override
-    public long getAverageGetMillis() {
-        return (long)(statistics.getAverageGetTimeMillis());
+    public float getAverageGetMillis() {
+        return (statistics.getAverageGetTimeMillis());
     }
 
     /**
@@ -196,7 +196,7 @@ public class JCacheStatistics implements javax.cache.CacheStatistics, Serializab
      * @return the time in milliseconds
      */
     @Override
-    public long getAveragePutMillis() {
+    public float getAveragePutMillis() {
         throw new UnsupportedOperationException("getAveragePutMillis is not implemented in net.sf.ehcache.jcache.JCacheStatistics");
     }
 
@@ -206,7 +206,7 @@ public class JCacheStatistics implements javax.cache.CacheStatistics, Serializab
      * @return the time in milliseconds
      */
     @Override
-    public long getAverageRemoveMillis() {
+    public float getAverageRemoveMillis() {
         throw new UnsupportedOperationException("getAverageRemoveMillis is not implemented in net.sf.ehcache.jcache.JCacheStatistics");
     }
 }
